% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aldvmm.check.R
\name{aldvmm.check}
\alias{aldvmm.check}
\title{Checking the Validity of Objects Supplied to
\ifelse{html}{\code{\link[aldvmm]{aldvmm}}}{\code{aldvmm::aldvmm()}}}
\usage{
aldvmm.check(
  formula,
  data,
  psi,
  ncmp,
  dist,
  lcoef,
  lcpar,
  lcmp,
  init.method,
  optim.method,
  optim.grad,
  init.est,
  init.lo,
  init.hi,
  optim.control,
  se.fit,
  level
)
}
\arguments{
\item{formula}{an object of class \code{"formula"} with a symbolic
description of the model to be fitted. The model formula takes the form
\code{y ~ x1 + x2 | x1 + x4}, where the \code{|} delimiter separates the
model for expected values of normal distributions (left) and the
multinomial logit model of probabilities of component membership (right)
(see details).}

\item{data}{a data frame, list or environment (or object coercible to a data
frame by \cr \ifelse{html}{\code{\link[base]{as.data.frame}}}{
\code{base::as.data.frame()}}) including data on outcomes and explanatory
variables in \code{'formula'}.}

\item{psi}{a numeric vector of minimum and maximum possible utility values
smaller than or equal to 1 (e.g. \code{c(-0.594, 0.883)}). The potential
gap between the maximum value and 1 represents an area with zero density
in the value set from which utilities were obtained. The order of the
minimum and maximum limits in \code{'psi'} does not matter.}

\item{ncmp}{a numeric value of the number of components that are mixed. The
default value is 2. A value of 1 represents a tobit model with a gap
between 1 and the maximum value in \code{'psi'}.}

\item{dist}{an optional character value of the distribution used in the
finite mixture. In this release, only the normal distribution is
available, and the default value is \code{"normal"}.}

\item{lcoef}{a character vector of length 2 with labels of objects including
regression coefficients. The first element is the label of coefficients
for component distributions (default \code{"beta"}), and the second
element is the label of coefficients on probabilities of component
membership (default \code{"delta"}).}

\item{lcpar}{a character vector with the labels of objects including
constant parameters of component distributions (e.g. the standard
deviation of the normal distribution). The length of \code{'lcpar'}
depends on the distribution supplied to \code{'dist'}.}

\item{lcmp}{a character value representing a stub (default \code{"Comp"})
for labeling objects including regression coefficients in different
components (e.g. "Comp1", "Comp2", ...). This label is also used in
summary tables returned by
\ifelse{html}{\code{\link[aldvmm]{summary.aldvmm}}}{
\code{aldvmm::summary.aldvmm()}}.}

\item{init.method}{an optional character value indicating the method for
obtaining initial values (see details). The following values are
available: \code{"zero"}, \code{"random"}, \code{"constant"} and
\code{"sann"}. The default value is \code{"zero"}.}

\item{optim.method}{an optional character value of one of the following
\ifelse{html}{\code{\link[optimr]{optimr}}}{\code{optimr::optimr()}}
methods: \code{"Nelder-Mead"}, \code{"BFGS"}, \code{"CG"},
\code{"L-BFGS-B"}, \code{"nlminb"}, \code{"Rcgmin"}, \code{"Rvmmin"} and
\code{"hjn"}. The default method is \code{"Nelder-Mead"}. The method
\code{"L-BFGS-B"} is used when lower and/or upper constraints are set
using \code{'init.lo'} and \code{'init.hi'}. The method \code{"nlm"}
cannot be used with the
\ifelse{html}{\code{\link[aldvmm]{aldvmm.ll}}}{\code{aldvm::aldvmm.ll()}}
likelihood function (see details).}

\item{optim.grad}{an optional logical value indicating if a numeric gradient
should be used in
\ifelse{html}{\code{\link[optimr]{optimr}}}{\code{optimr::optimr()}}
methods that can use this information. If \code{'optim.grad'} is
\code{FALSE}, a finite difference approximation will be used.}

\item{init.est}{an optional numeric vector of user-defined initial values.
User-defined initial values override the \code{'init.method'} argument.
Initial values have to follow the same order as parameter estimates appear
in the return value \code{'par'}.}

\item{init.lo}{an optional numeric vector of user-defined lower limits for
constrained optimization. When \code{'init.lo'} is not \code{NULL}, the
optimization method \code{"L-BFGS-B"} is used. Lower limits of parameters
have to follow the same order as parameter estimates appear in the return
value \code{'par'}.}

\item{init.hi}{an optional numeric vector of user-defined upper limits for
constrained optimization. When \code{'init.hi'} is not \code{NULL}, the
optimization method \code{"L-BFGS-B"} is used. Upper limits of parameters
have to follow the same order as parameter estimates appear in the return
value \code{'par'}.}

\item{optim.control}{an optional list of
\ifelse{html}{\code{\link[optimr]{optimr}}}{\code{optimr::optimr()}}
control parameters.}

\item{se.fit}{an optional logical value indicating whether standard errors
of the fit are calculated (see Details).}

\item{level}{a numeric value of the significance level for confidence bands
of fitted values. The default is 0.95.}
}
\value{
\code{aldvmm.check} returns warnings or stops the execution of
  \ifelse{html}{\code{\link[aldvmm]{aldvmm}}}{\code{aldvmm::aldvmm()}} if
  validity checks fail.
}
\description{
\ifelse{html}{\code{\link[aldvmm]{aldvmm.check}}}{\code{aldvmm::aldvmm.check()}}
runs validity checks of objects supplied to
\ifelse{html}{\code{\link[aldvmm]{aldvmm}}}{\code{aldvmm::aldvmm()}}.
}
\details{
\ifelse{html}{\code{\link[aldvmm]{aldvmm.check}}}{\code{aldvmm::aldvmm.check()}}
  checks the validity of input values of
  \ifelse{html}{\code{\link[aldvmm]{aldvmm}}}{\code{aldvmm::aldvmm()}} and
  of user settings.
}
