% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aldvmm.gr.R
\name{aldvmm.gr}
\alias{aldvmm.gr}
\title{Calculating analytical Gradients of the Negative Log-Likelihood}
\usage{
aldvmm.gr(
  par,
  X,
  y,
  psi,
  dist,
  ncmp,
  lcoef = lcoef,
  lcmp = lcmp,
  lcpar = lcpar
)
}
\arguments{
\item{sum.grad}{a logical value indicating whether the sum of gradients of
all observations for the likelihood of the full data is returned 
(\code{TRUE}) or a matrix of gradients per observation (\code{FALSE})}
}
\value{
a named numeric matrix of first derivatives of the negative
  log-likelihood of the data with respect to parameters in \code{'par'}.
}
\description{
\ifelse{html}{\code{\link[aldvmm]{aldvmm.gr}}}{\code{aldvmm.gr()}}
calculates analytical gradients of the negative log-likelihood with respect 
to parameter values in \code{'par'} for each observation in the estimation 
data.
}
\details{
\ifelse{html}{\code{\link[aldvmm]{aldvmm.gr}}}{\code{aldvmm.gr()}}
calculates gradients of the negative log-likelihood.
}
