% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aldvmm.sum.R
\name{aldvmm.sum}
\alias{aldvmm.sum}
\title{Creating Summary Table}
\usage{
aldvmm.sum(
  est,
  se,
  z,
  p,
  lower,
  upper,
  n,
  value,
  aic,
  bic,
  ncmp,
  lcoef,
  lcpar,
  lcmp,
  lvar
)
}
\arguments{
\item{est}{a named numeric vector of parameters with names returned by
\ifelse{html}{\code{\link[aldvmm]{aldvmm.getnames}}}{
\code{aldvmm::aldvmm.getnames()}}.}

\item{se}{a named numeric vector of standard errors of parameters returned
by \ifelse{html}{\code{\link[aldvmm]{aldvmm.cv}}}{
\code{aldvmm::aldvmm.cv()}}.}

\item{z}{a named numeric vector of z scores (standardized coefficients) of
parameters returned by \ifelse{html}{\code{\link[aldvmm]{aldvmm.cv}}}{
\code{aldvmm::aldvmm.cv()}}.}

\item{p}{a named numeric vector of p-values of parameters returned by
\ifelse{html}{\code{\link[aldvmm]{aldvmm.cv}}}{
\code{aldvmm::aldvmm.cv()}}.}

\item{lower}{a named numeric vector of 95\% lower limits of parameters
returned by \ifelse{html}{\code{\link[aldvmm]{aldvmm.cv}}}{
\code{aldvmm::aldvmm.cv()}}.}

\item{upper}{a named numeric vector of 95\% upper limits of parameters
returned by \ifelse{html}{\code{\link[aldvmm]{aldvmm.cv}}}{
\code{aldvmm::aldvmm.cv()}}.}

\item{n}{a numeric value of the number of complete observations in
\code{'data'} supplied to
\ifelse{html}{\code{\link[aldvmm]{aldvmm}}}{\code{aldvmm::aldvmm()}}.}

\item{value}{a numeric value of the negative log-likelihood returned by
\ifelse{html}{\code{\link[aldvmm]{aldvmm.ll}}}{
\code{aldvmm::aldvmm.ll()}}.}

\item{aic}{a numeric value of the Akaike information criterion (AIC)
returned by \ifelse{html}{\code{\link[aldvmm]{aldvmm}}}{
\code{aldvmm::aldvmm()}}.}

\item{bic}{a numeric value of the Bayesian information criterion (BIC)
returned by \ifelse{html}{\code{\link[aldvmm]{aldvmm}}}{
\code{aldvmm::aldvmm()}}.}

\item{ncmp}{a numeric value of the number of components that are mixed. The
default value is 2. A value of 1 represents a tobit model with a gap
between 1 and the maximum value in \code{'psi'}.}

\item{lcoef}{a character vector of the names of objects with values for
component distributions (default \code{"beta"}) and for probabilities of
component membership (default \code{"delta"}). The order in which the
labels in the two parts appear affects ordering of results in parameter
vectors and summary tables. The names in \code{'lcoef'} are allowed to
coincide with column names in \code{'data'} supplied to
\ifelse{html}{\code{\link[aldvmm]{aldvmm}}}{\code{aldvmm::aldvmm()}}.}

\item{lcpar}{a character vector of the names of objects with values of
constant parameters of component distributions. The names of constant
parameters depend on the chosen distribution (\code{"lnsigma"} for
\code{dist = "normal"}). The names in \code{'lcpar'} are allowed to
coincide with column names in \code{'data'} supplied to
\ifelse{html}{\code{\link[aldvmm]{aldvmm}}}{\code{aldvmm::aldvmm()}}.}

\item{lcmp}{a character value of the names of objects with values for a
specific component (default \code{"Comp"}). This label is also used in
summary tables returned by
\ifelse{html}{\code{\link[aldvmm]{summary.aldvmm}}}{
\code{aldvmm::summary.aldvmm()}}.}

\item{lvar}{a named list of character vectors with column names of design
matrices returned by \ifelse{html}{\code{\link[aldvmm]{aldvmm.mm}}}{
\code{aldvmm::aldvmm.mm()}}.}
}
\value{
a data frame with a summary table of regression results.
}
\description{
\ifelse{html}{\code{\link[aldvmm]{aldvmm.sum}}}{
\code{aldvmm::aldvmm.sum()}} creates a summary table of regression results.
}
