% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{formula.aldvmm}
\alias{formula.aldvmm}
\title{Extract Adjusted Limited Dependent Variable Mixture Model Formula}
\usage{
\method{formula}{aldvmm}(object, ...)
}
\arguments{
\item{object}{an object inheriting from class "aldvmm".}

\item{...}{further arguments passed to or from other methods.}
}
\value{
an object of class "formula"
}
\description{
The method \code{formula.aldvmm} for the generic function
\ifelse{html}{\code{\link[stats]{formula}}}{\code{stats::formula()}} returns the 
formula object from an object of class "aldvmm".
}
