% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.matrix.aldvmm.R
\name{model.matrix.aldvmm}
\alias{model.matrix.aldvmm}
\title{Extract Adjusted Limited Dependent Variable Mixture Model Model Matrices}
\usage{
\method{model.matrix}{aldvmm}(object, ...)
}
\arguments{
\item{object}{an object inheriting from class "aldvmm".}

\item{...}{further arguments passed to or from other methods.}
}
\description{
The method
\code{model.matrix.aldvmm} extracts a list of model matrices from an object 
of class "aldvmm" using the function 
\ifelse{html}{\code{\link[aldvmm]{aldvmm.mm}}}{\code{aldvmm::aldvmm.mm()}}.
}
