% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nobs.aldvmm.R
\name{nobs.aldvmm}
\alias{nobs.aldvmm}
\title{Obtain Adjusted Limited Dependent Variable Mixture Model Number of Observations}
\usage{
\method{nobs}{aldvmm}(object, ...)
}
\arguments{
\item{object}{an object inheriting from class "aldvmm".}

\item{...}{further arguments passed to or from other methods.}
}
\description{
The generic function
\ifelse{html}{\code{\link[stats]{nobs}}}{\code{stats::nobs()}} extracts the 
number of observations from an object of class "aldvmm".
}
