% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.aldvmm}
\alias{print.aldvmm}
\title{Print Adjusted Limited Dependent Variable Mixture Model Fits}
\usage{
\method{print}{aldvmm}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{an object inheriting from class "aldvmm".}

\item{digits}{an integer value of the number of digits in the printed 
output.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
The method \code{print.aldvmm} for the generic function
\ifelse{html}{\code{\link[base]{print}}}{\code{base::print()}} prints a 
summary of an object of class "aldvmm".
}
