% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.aldvmm.R
\name{print.summary.aldvmm}
\alias{print.summary.aldvmm}
\title{Printing Adjusted Limited Dependent Variable Mixture Model Summaries}
\usage{
\method{print}{summary.aldvmm}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{an object inheriting from class 'summary.aldvmm'.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
The method
\code{print.summary.aldvmm} prints a summary of an 
\ifelse{html}{\code{\link[aldvmm]{aldvmm}}}{\code{aldvmm::aldvmm()}} model 
fit.
}
