% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.aldvmm.R
\name{residuals.aldvmm}
\alias{residuals.aldvmm}
\title{Extract Adjusted Limited Dependent Variable Mixture Model Residuals}
\usage{
\method{residuals}{aldvmm}(object, ...)
}
\arguments{
\item{object}{an object inheriting from class "aldvmm".}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a numeric vector of residuals.
}
\description{
The method \code{residuals.aldvmm} for the generic function
\ifelse{html}{\code{\link[stats]{residuals}}}{\code{stats::residuals()}} returns the 
covariance matrix from an object of class "aldvmm".
}
