% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.aldvmm.R
\name{summary.aldvmm}
\alias{summary.aldvmm}
\title{Summarizing Adjusted Limited Dependent Variable Mixture Model Fits}
\usage{
\method{summary}{aldvmm}(object, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{object}{an object inheriting from class 'aldvmm'.}

\item{digits}{an integer value of the number of digits in the output table.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
\code{summary.aldvmm} returns an object of class "summary.aldvmm" 
including the following elements. 
  \item{\code{call}}{a character value including the model call captured by
  \ifelse{html}{\code{\link[base]{match.call}}}{\code{base::match.call}}.}
  \item{\code{summary}}{a data frame generated by 
  \ifelse{html}{\code{\link[aldvmm]{aldvmm.sum}}}{\code{aldvmm::aldvmm.sum}} 
  including a formatted summary of model results for printing.}
  \item{\code{terms}}{a list of objects of class
  \ifelse{html}{\code{\link[stats]{terms}}}{\code{stats::terms}}.}
  \item{\code{contrasts}}{a nested list of character values showing 
  contrasts of factors used in models of component means ("beta") and 
  probabilities of component membership ("delta").}
  \item{\code{coef}}{a
  numeric vector of parameter estimates.}
  \item{\code{cov}}{a numeric matrix object with covariances of parameters.}
  \item{\code{se}}{a numeric vector of standard errors of parameter estimates.}
  \item{\code{z}}{a numeric vector of standardized parameter estimates.}
  \item{\code{lower}}{a numeric
  vector of 95\% lower confidence limits of parameter estimates.}
  \item{\code{upper}}{a numeric vector of 95\% upper confidence limits of
  parameter estimates.}
  \item{\code{n}}{a scalar representing the number of complete observations
  with no missing values that were used in the estimation.}
  \item{\code{df.residual}}{a list including integer value of the residual 
  degrees of freedom for the model of component means ("beta") and 
  probability of component membership ("delta").}
  \item{\code{df.null}}{a list including integer value of the residual 
  degrees of freedom of a null model of component means ("beta") and 
  probability of component membership ("delta").}
  \item{\code{iter}}{an integer value of the number of iterations used in 
  optimization.}
  \item{\code{ll}}{a numeric value of the negative log-likelihood
  \eqn{-ll}.}
  \item{\code{aic}}{a numeric value of the Akaike information
  criterion \eqn{AIC = 2n_{par} - 2ll}{AIC = 2*npar - 2*ll}.}
  \item{\code{bic}}{a numeric value of the Bayesian information criterion
  \eqn{BIC = n_{par}*log(n_{obs}) - 2ll}{BIC = npar*log(nobs) - 2*ll}.}
  \item{\code{lcoef}}{a character vector of labels for objects including
  results on distributions (default \code{"beta"}) and the probabilities of
  component membership (default \code{"delta"}).} 
  \item{\code{lcpar}}{a
  character vector of labels for objects including constant distribution
  parameters (default \code{"sigma"} for \code{dist = "normal"}).}
  \item{\code{lcmp}}{a character value of the label for objects including
  results on different components (default "Comp")} 
  \item{\code{lvar}}{a
  list including 2 character vectors of covariate names for model parameters
  of distributions (\code{"beta"}) and the multinomial logit
  (\code{"delta"}).}
}
\description{
The generic function
\ifelse{html}{\code{\link[base]{summary}}}{\code{base::summary()}} creates 
an object of class "summary.aldvmm" including key results from an 
\ifelse{html}{\code{\link[aldvmm]{aldvmm}}}{\code{aldvmm::aldvmm()}} model
fit.
}
