% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aldvmmMethods.R
\name{terms.aldvmm}
\alias{terms.aldvmm}
\title{Extract Adjusted Limited Dependent Variable Mixture Model Terms}
\usage{
\method{terms}{aldvmm}(object, ...)
}
\arguments{
\item{object}{an object inheriting from class "aldvmm".}

\item{...}{further arguments passed to or from other methods.}
}
\value{
an object of class "terms"
}
\description{
The method \code{terms.aldvmm} for the generic function
\ifelse{html}{\code{\link[stats]{terms}}}{\code{stats::terms()}} returns the 
terms object for the combined model of component means and probabilities of
component membership from an object of class "aldvmm".
}
