% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{update.aldvmm}
\alias{update.aldvmm}
\title{Update Adjusted Limited Dependent Variable Mixture Model Fit}
\usage{
\method{update}{aldvmm}(object, formula., ..., evaluate = TRUE)
}
\arguments{
\item{object}{an object inheriting from class "aldvmm".}

\item{formula.}{a formula object representing the new model.}

\item{...}{further arguments passed to or from other methods.}

\item{evaluate}{a logical value indicating if the model should be 
re-estimated (TRUE) or not (FALSE).}
}
\value{
an object of class "aldvmm"
}
\description{
The method \code{update.aldvmm} for the generic function
\ifelse{html}{\code{\link[stats]{update}}}{\code{stats::update()}} 
re-estimates an object of class "aldvmm".
}
