% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{vcov.aldvmm}
\alias{vcov.aldvmm}
\title{Extract Adjusted Limited Dependent Variable Mixture Model Covariance Matrix}
\usage{
\method{vcov}{aldvmm}(object, ...)
}
\arguments{
\item{object}{an object inheriting from class "aldvmm".}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a numeric matrix.
}
\description{
The method \code{vcov.aldvmm} for the generic function
\ifelse{html}{\code{\link[stats]{vcov}}}{\code{stats::vcov()}} returns the 
covariance matrix from an object of class "aldvmm".
}
