% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aldvmm.cv.R
\name{aldvmm.cv}
\alias{aldvmm.cv}
\title{Numerical Approximation of Covariance Matrix}
\usage{
aldvmm.cv(ll, par, X, y, dist, psi, ncmp, lcoef, lcpar, lcmp, optim.method)
}
\arguments{
\item{ll}{a function
\ifelse{html}{\code{\link[aldvmm]{aldvmm.ll}}}{\code{aldvmm::aldvmm.ll()}}
returning the negative log-likelihood of the adjusted limited dependent
variable mixture model as a scalar result.}

\item{par}{a named numeric vector of parameter values used to calculate the
log-likelihood. The names of the parameter vector must be generated using
\ifelse{html}{\code{\link[aldvmm]{aldvmm.getnames}}}{
\code{aldvmm::aldvmm.getnames()}} because they will be inherited by return
values of functions in package \code{'aldvmm'}, and the names will be used
in the extraction of parameters into nested lists using \ifelse{html}{
\code{\link[aldvmm]{aldvmm.getpar}}}{\code{aldvmm::aldvmm.getpar()}}.}

\item{X}{a list of design matrices returned by
\ifelse{html}{\code{\link[aldvmm]{aldvmm.mm}}}{\code{aldvmm::aldvmm.mm()}}.
 \code{'X'} is of length 2 and includes numeric design matrices for the
models of component distributions (label \code{"beta"}) and the
probabilities of group membership (label \code{"delta"}). The numeric
matrices in \code{'X'} have rownames inherited from \code{'data'} supplied
to \ifelse{html}{\code{\link[aldvmm]{aldvmm}}}{\code{aldvmm::aldvmm()}}.
The rownames will be used in
\ifelse{html}{\code{\link[aldvmm]{aldvmm.pred}}}{
\code{aldvmm::aldvmm.pred()}} to return missing predictions for incomplete
observations in \code{'data'}.}

\item{y}{a numeric vector of observed outcomes from complete observations in
\code{'data'} supplied to
\ifelse{html}{\code{\link[aldvmm]{aldvmm}}}{\code{aldvmm::aldvmm()}}.}

\item{dist}{an optional character value of the distribution used in the
finite mixture. In this release, only the normal distribution is
available, and the default value is \code{"normal"}.}

\item{psi}{a numeric vector of minimum and maximum possible utility values
smaller than or equal to 1 (e.g. \code{c(-0.594, 0.883)}). The potential
gap between the maximum value and 1 represents an area with zero density
in the value set from which utilities were obtained. The order of the
minimum and maximum limits in \code{'psi'} does not matter.}

\item{ncmp}{a numeric value of the number of components that are mixed. The
default value is 2. A value of 1 represents a tobit model with a gap
between 1 and the maximum value in \code{'psi'}.}

\item{lcoef}{a character vector of the names of objects with values for
component distributions (default \code{"beta"}) and for probabilities of
component membership (default \code{"delta"}). The order in which the
labels in the two parts appear affects ordering of results in parameter
vectors and summary tables. The names in \code{'lcoef'} are allowed to
coincide with column names in \code{'data'} supplied to
\ifelse{html}{\code{\link[aldvmm]{aldvmm}}}{\code{aldvmm::aldvmm()}}.}

\item{lcpar}{a character vector of the names of objects with values of
constant parameters of component distributions. The names of constant
parameters depend on the chosen distribution (\code{"lnsigma"} for
\code{dist = "normal"}). The names in \code{'lcpar'} are allowed to
coincide with column names in \code{'data'} supplied to
\ifelse{html}{\code{\link[aldvmm]{aldvmm}}}{\code{aldvmm::aldvmm()}}.}

\item{lcmp}{a character value of the names of objects with values for a
specific component (default \code{"Comp"}). This label is also used in
summary tables returned by
\ifelse{html}{\code{\link[aldvmm]{summary.aldvmm}}}{
\code{aldvmm::summary.aldvmm()}}.}

\item{optim.method}{an optional character value of one of the following
\ifelse{html}{\code{\link[optimr]{optimr}}}{\code{optimr::optimr()}}
methods: \code{"Nelder-Mead"}, \code{"BFGS"}, \code{"CG"},
\code{"L-BFGS-B"}, \code{"nlminb"}, \code{"Rcgmin"}, \code{"Rvmmin"} and
\code{"hjn"}. The default method is \code{"Nelder-Mead"}. The method
\code{"L-BFGS-B"} is used when lower and/or upper constraints are set
using \code{'init.lo'} and \code{'init.hi'}. The method \code{"nlm"}
cannot be used with the
\ifelse{html}{\code{\link[aldvmm]{aldvmm.ll}}}{\code{aldvm::aldvmm.ll()}}
likelihood function (see details).}
}
\value{
\code{aldvmm.cv} returns a list with the following objects.
  \item{\code{hessian}}{a numeric matrix with second-order partial
  derivatives of the likelihood function \code{'ll'}.} \item{\code{cv}}{a
  numeric matrix with covariances/variances of parameters in \code{'par'}.}
  \item{\code{se}}{a numeric vector of standard errors of parameters in
  \code{'par'}.} \item{\code{z}}{a numeric vector of z-values (parameter
  values in \code{'par'} divided by standard error \code{'se'}).}
  \item{\code{p}}{a numeric vector of p-values of parameter estimates
  (probabilities of \code{'z'} smaller than zero).} \item{\code{upper}}{a
  numeric vector of upper 95\% confindence limits of paramter estimates in
  \code{'par'}.} \item{\code{lower}}{a numeric vector of lower 95\%
  confindence limits of paramter estimates in \code{'par'}.}
}
\description{
\ifelse{html}{\code{\link[aldvmm]{aldvmm.cv}}}{\code{aldvmm::aldvmm.cv()}}
performs a numerical approximation of the covariance matrix of parameter
estimates.
}
\details{
\ifelse{html}{\code{\link[aldvmm]{aldvmm.cv}}}{\code{aldvmm::aldvmm.cv()}}
uses
\ifelse{html}{\code{\link[numDeriv]{hessian}}}{\code{numDeriv::hessian()}}
to calculate the hessian matrix of the log-likelihood function in
\code{'ll'} at parameter values in \code{'par'}.
}
\author{
Mark Pletscher, <pletscher.mark@gmail.com>
}
