% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aldvmm.getnames.R
\name{aldvmm.getnames}
\alias{aldvmm.getnames}
\title{Creating Names of Parameter Vectors}
\usage{
aldvmm.getnames(X, names, lcoef, lcpar, lcmp, ncmp)
}
\arguments{
\item{X}{a list of design matrices returned by
\ifelse{html}{\code{\link[aldvmm]{aldvmm.mm}}}{\code{aldvmm::aldvmm.mm()}}.
 \code{'X'} is of length 2 and includes numeric design matrices for the
models of component distributions (label \code{"beta"}) and the
probabilities of group membership (label \code{"delta"}). The numeric
matrices in \code{'X'} have rownames inherited from \code{'data'} supplied
to \ifelse{html}{\code{\link[aldvmm]{aldvmm}}}{\code{aldvmm::aldvmm()}}.
The rownames will be used in
\ifelse{html}{\code{\link[aldvmm]{aldvmm.pred}}}{
\code{aldvmm::aldvmm.pred()}} to return missing predictions for incomplete
observations in \code{'data'}.}

\item{names}{a character vector of names of considered elements
(distributions \code{"beta"}, multinomial logit \code{"delta"} or constant
distribution parameters, i.e. \code{"lnsigma"}). The elements in
\code{'names'} are combined with covariate names in \code{'X'} and
component labels in \code{'lcmp'} to create a vector of names of parameter
vectors.}

\item{lcoef}{a character vector of the names of objects with values for
component distributions (default \code{"beta"}) and for probabilities of
component membership (default \code{"delta"}). The order in which the
labels in the two parts appear affects ordering of results in parameter
vectors and summary tables. The names in \code{'lcoef'} are allowed to
coincide with column names in \code{'data'} supplied to
\ifelse{html}{\code{\link[aldvmm]{aldvmm}}}{\code{aldvmm::aldvmm()}}.}

\item{lcpar}{a character vector of the names of objects with values of
constant parameters of component distributions. The names of constant
parameters depend on the chosen distribution (\code{"lnsigma"} for
\code{dist = "normal"}). The names in \code{'lcpar'} are allowed to
coincide with column names in \code{'data'} supplied to
\ifelse{html}{\code{\link[aldvmm]{aldvmm}}}{\code{aldvmm::aldvmm()}}.}

\item{lcmp}{a character value of the names of objects with values for a
specific component (default \code{"Comp"}). This label is also used in
summary tables returned by
\ifelse{html}{\code{\link[aldvmm]{summary.aldvmm}}}{
\code{aldvmm::summary.aldvmm()}}.}

\item{ncmp}{a numeric value of the number of components that are mixed. The
default value is 2. A value of 1 represents a tobit model with a gap
between 1 and the maximum value in \code{'psi'}.}
}
\value{
a character vector of names of parameter vectors used in
  \ifelse{html}{\code{\link[aldvmm]{aldvmm}}}{\code{aldvmm::aldvmm()}}.
}
\description{
\ifelse{html}{\code{\link[aldvmm]{aldvmm.getnames}}}{\code{aldvmm::aldvmm.getnames()}}
creates names of parameter vectors used in
\ifelse{html}{\code{\link[aldvmm]{aldvmm}}}{\code{aldvmm::aldvmm()}}. The
order of the elements in \code{'lcoef'} and \code{'lcpar'} determines the
order of parameters and the structure of summary tables returned by
\ifelse{html}{\code{\link[aldvmm]{summary.aldvmm}}}{\code{aldvmm::summary.aldvmm()}}.
}
