% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aldvmm.getpar.R
\name{aldvmm.getpar}
\alias{aldvmm.getpar}
\title{Extracting Parameters from Parameter Vector into Nested List.}
\usage{
aldvmm.getpar(par, lcoef, lcmp, lcpar, ncmp)
}
\arguments{
\item{par}{a named numeric vector of parameter values used to calculate the
log-likelihood. The names of the parameter vector must be generated using
\ifelse{html}{\code{\link[aldvmm]{aldvmm.getnames}}}{
\code{aldvmm::aldvmm.getnames()}} because they will be inherited by return
values of functions in package \code{'aldvmm'}, and the names will be used
in the extraction of parameters into nested lists using \ifelse{html}{
\code{\link[aldvmm]{aldvmm.getpar}}}{\code{aldvmm::aldvmm.getpar()}}.}

\item{lcoef}{a character vector of the names of objects with values for
component distributions (default \code{"beta"}) and for probabilities of
component membership (default \code{"delta"}). The order in which the
labels in the two parts appear affects ordering of results in parameter
vectors and summary tables. The names in \code{'lcoef'} are allowed to
coincide with column names in \code{'data'} supplied to
\ifelse{html}{\code{\link[aldvmm]{aldvmm}}}{\code{aldvmm::aldvmm()}}.}

\item{lcmp}{a character value of the names of objects with values for a
specific component (default \code{"Comp"}). This label is also used in
summary tables returned by
\ifelse{html}{\code{\link[aldvmm]{summary.aldvmm}}}{
\code{aldvmm::summary.aldvmm()}}.}

\item{lcpar}{a character vector of the names of objects with values of
constant parameters of component distributions. The names of constant
parameters depend on the chosen distribution (\code{"lnsigma"} for
\code{dist = "normal"}). The names in \code{'lcpar'} are allowed to
coincide with column names in \code{'data'} supplied to
\ifelse{html}{\code{\link[aldvmm]{aldvmm}}}{\code{aldvmm::aldvmm()}}.}

\item{ncmp}{a numeric value of the number of components that are mixed. The
default value is 2. A value of 1 represents a tobit model with a gap
between 1 and the maximum value in \code{'psi'}.}
}
\value{
a named nested list with parameter vectors for \code{"beta"},
  \code{"delta"} and \code{"lnsigma"} within all \code{1:K} components. The
  names of the list correspond to the labels in \code{'lcoef'},
  \code{'lcpar'} and \code{'lcmp'}.
}
\description{
\code{aldvmm.getpar} extracts parameters from parameter vectors into nested
lists.
}
\details{
\ifelse{html}{\code{\link[aldvmm]{aldvmm.getpar}}}{\code{aldvmm::aldvmm.getpar()}}
  identifies parameters that belong to coefficients of component
  distributions (label \code{"beta"}), coefficients of the multinomial logit
  model of probabilities of component membership (label \code{"delta"}),
  constant parameters of component distributions (label \code{"lnsigma"} for
  \code{dist=="normal"} and \code{1:K} components (labels "Comp1" ...
  "CompK")) based on parameter names generated by
  \ifelse{html}{\code{\link[aldvmm]{aldvmm.getnames}}}{\code{aldvmm::aldvmm.getnames()}}.
}
