% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.aldvmm.R
\name{coef.aldvmm}
\alias{coef.aldvmm}
\title{Extract Model Coefficients of Adjusted Limited Dependent Variable Mixture Model Fits}
\usage{
\method{coef}{aldvmm}(object, ...)
}
\arguments{
\item{object}{an object inheriting from class 'aldvmm'.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a named vector of numeric parameter estimates.
}
\description{

}
