% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estfun.aldvmm.R
\name{estfun.aldvmm}
\alias{estfun.aldvmm}
\title{Create Matrix of Adjusted Limited Dependent Variable Mixture Model Gradients per Observation}
\usage{
\method{estfun}{aldvmm}(object, ...)
}
\arguments{
\item{object}{an object inheriting from class "aldvmm".}

\item{...}{further arguments passed to or from other methods.}
}
\value{
\code{estfun.aldvmm} returns a matrix with one row per observation and one column per parameter.
}
\description{
The generic function 
\ifelse{html}{\code{\link[sandwich]{estfun}}}{\code{sandwich::estfun()}} calculates the gradient of the aldvmm 
log-likelihood 
\ifelse{html}{\code{\link[aldvmm]{aldvmm.ll}}}{\code{aldvmm::aldvmm.ll()}} 
with respect to parameter values for each observation using 
\ifelse{html}{\code{\link[numDeriv]{jacobian}}}{\code{numDeriv::jacobian()}}.
}
