% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estfun.aldvmm.R
\name{estfun.aldvmm}
\alias{estfun.aldvmm}
\title{Create Matrix of Adjusted Limited Dependent Variable Mixture Model Gradients per Observation}
\usage{
\method{estfun}{aldvmm}(x, ...)
}
\arguments{
\item{x}{an object inheriting from class "aldvmm".}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a numeric matrix of gradients with one row per 
observation and one column per parameter.
}
\description{
The method \code{estfun.aldvmm} for the generic function 
\ifelse{html}{\code{\link[sandwich]{estfun}}}{\code{sandwich::estfun()}} calculates the gradient of the aldvmm 
log-likelihood using with respect to parameter values for each observation
\ifelse{html}{\code{\link[aldvmm]{aldvmm.sc}}}{\code{aldvmm::aldvmm.sc()}}.
}
