% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aldvmmMethods.R
\name{model.matrix.aldvmm}
\alias{model.matrix.aldvmm}
\title{Extract Adjusted Limited Dependent Variable Mixture Model Model Matrices}
\usage{
\method{model.matrix}{aldvmm}(object, ...)
}
\arguments{
\item{object}{an object inheriting from class "aldvmm".}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a named list of numeric matrices. 
\item{\code{beta}}{a numeric design matrix for the model of component 
distributions.}
\item{\code{delta}}{a numeric design matrix of the multinomial logit model
  of probabilities of component membership.}
}
\description{
The method
\code{model.matrix.aldvmm} for the generic function \ifelse{html}{\code{\link[stats]{model.matrix}}}{\code{stats::model.matrix()}} extracts a list of model matrices from an object 
of class "aldvmm" using the function 
\ifelse{html}{\code{\link[aldvmm]{aldvmm.mm}}}{\code{aldvmm::aldvmm.mm()}}.
}
