% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.aldvmm.R
\name{print.aldvmm}
\alias{print.aldvmm}
\title{Print Adjusted Limited Dependent Variable Mixture Model Fits}
\usage{
\method{print}{aldvmm}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{an object inheriting from class "aldvmm".}

\item{digits}{an integer value of the number of digits in the printed 
output.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
The generic function
\ifelse{html}{\code{\link[stats]{coef}}}{\code{stats::print()}} prints key 
attributes of an object of class "aldvmm".
}
