% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.summary.aldvmm}
\alias{print.summary.aldvmm}
\title{Printing Adjusted Limited Dependent Variable Mixture Model Summaries}
\usage{
\method{print}{summary.aldvmm}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{an object inheriting from class 'summary.aldvmm'.}

\item{digits}{an integer value of the number of digits in the output table.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
The method \code{print.summary.aldvmm} for the generic function
\ifelse{html}{\code{\link[base]{print}}}{\code{base::print()}} prints a 
summary of an object of class "summary.aldvmm".
}
